Shifting of the aperture position
If the beam is inserted elsewhere in sixtrack, the aperture needs to be
shifted. Inserting in front of the first collimator (in the
inparameters.dat list) is now possible. If the new value is 0, no
shifting is done, if it is the position of the collimator it will be
shifted by that value